airflow_project/requirements.txt
requests
beautifulsoup4
lxml
pandas
google-generativeai
python-dotenv